Refactor tests: combine withdraw tests, separate signature tests
